+++ packages/kit/src/runtime/client/fetcher.js
@@ -53,7 +53,7 @@ if (DEV && BROWSER) {
 
 		if (in_load_heuristic && !used_kit_fetch) {
 			console.warn(
-				`Loading ${url} using \`window.fetch\`. For best results, use the \`fetch\` that is passed to your \`load\` function: https://kit.svelte.dev/docs/load#making-fetch-requests`
+				`Loading ${url} using \`window.fetch\`. For best results, use the \`fetch\` that is passed to your \`load\` function: https://svelte.dev/docs/kit/load#making-fetch-requests`
 			);
 		}