+++ packages/kit/src/runtime/server/page/actions.js
@@ -197,7 +197,7 @@ export async function handle_action_request(event, server) {
 function check_named_default_separate(actions) {
 	if (actions.default && Object.keys(actions).length > 1) {
 		throw new Error(
-			'When using named actions, the default action cannot be used. See the docs for more info: https://kit.svelte.dev/docs/form-actions#named-actions'
+			'When using named actions, the default action cannot be used. See the docs for more info: https://svelte.dev/docs/kit/form-actions#named-actions'
 		);
 	}
 }
@@ -280,6 +280,14 @@ function try_deserialize(data, fn, route_id) {
 		// If we're here, the data could not be serialized with devalue
 		const error = /** @type {any} */ (e);
 
+		// if someone tries to use `json()` in their action
+		if (data instanceof Response) {
+			throw new Error(
+				`Data returned from action inside ${route_id} is not serializable. Form actions need to return plain objects or fail(). E.g. return { success: true } or return fail(400, { message: "invalid" });`
+			);
+		}
+
+		// if devalue could not serialize a property on the object, etc.
 		if ('path' in error) {
 			let message = `Data returned from action inside ${route_id} is not serializable: ${error.message}`;
 			if (error.path !== '') message += ` (data.${error.path})`;