+++ packages/kit/src/utils/url.spec.js
@@ -102,7 +102,7 @@ describe('make_trackable', (test) => {
 	test('makes URL properties trackable', () => {
 		let tracked = false;
 		const url = make_trackable(
-			new URL('https://kit.svelte.dev/docs'),
+			new URL('https://svelte.dev/docs/kit'),
 			() => {
 				tracked = true;
 			},
@@ -118,7 +118,7 @@ describe('make_trackable', (test) => {
 
 	test('throws an error when its hash property is accessed', () => {
 		const url = make_trackable(
-			new URL('https://kit.svelte.dev/docs'),
+			new URL('https://svelte.dev/docs/kit'),
 			() => {},
 			() => {}
 		);
@@ -133,7 +133,7 @@ describe('make_trackable', (test) => {
 		let tracked = false;
 		const tracked_search_params = new Set();
 		const url = make_trackable(
-			new URL('https://kit.svelte.dev/docs'),
+			new URL('https://svelte.dev/docs/kit'),
 			() => {
 				tracked = true;
 			},
@@ -161,7 +161,7 @@ describe('make_trackable', (test) => {
 
 describe('disable_search', (test) => {
 	test('throws an error when its search property is accessed', () => {
-		const url = new URL('https://kit.svelte.dev/docs');
+		const url = new URL('https://svelte.dev/docs/kit');
 		disable_search(url);
 
 		/** @type {Array<keyof URL>} */