+++ packages/migrate/migrations/sveltekit-2/migrate.js
@@ -52,7 +52,7 @@ export function update_tsconfig_content(content) {
 	if (updated !== content) {
 		log_migration(
 			'Removed deprecated `importsNotUsedAsValues` and `preserveValueImports`' +
-				' from tsconfig.json: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#updated-dependency-requirements'
+				' from tsconfig.json: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#updated-dependency-requirements'
 		);
 	}
 
@@ -61,13 +61,13 @@ export function update_tsconfig_content(content) {
 	if (updated !== content) {
 		log_migration(
 			'Updated `moduleResolution` to `bundler`' +
-				' in tsconfig.json: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#updated-dependency-requirements'
+				' in tsconfig.json: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#updated-dependency-requirements'
 		);
 	}
 
 	if (content.includes('"paths":') || content.includes('"baseUrl":')) {
 		log_migration(
-			'`paths` and/or `baseUrl` detected in your tsconfig.json - remove it and use `kit.alias` instead: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#generated-tsconfig-json-is-more-strict'
+			'`paths` and/or `baseUrl` detected in your tsconfig.json - remove it and use `kit.alias` instead: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#generated-tsconfig-json-is-more-strict'
 		);
 	}
 
@@ -89,7 +89,7 @@ export function update_svelte_config_content(code) {
 	const result = code.replace(regex, '');
 	if (result !== code) {
 		log_migration(
-			'Removed `dangerZone` from svelte.config.js: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#server-fetches-are-not-trackable-anymore'
+			'Removed `dangerZone` from svelte.config.js: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#server-fetches-are-not-trackable-anymore'
 		);
 	}
 
@@ -101,7 +101,7 @@ export function update_svelte_config_content(code) {
 
 	const logger = log_on_ts_modification(
 		source,
-		'Changed `vitePreprocess` import: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#vitepreprocess-is-no-longer-exported-from-sveltejs-kit-vite'
+		'Changed `vitePreprocess` import: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#vitepreprocess-is-no-longer-exported-from-sveltejs-kit-vite'
 	);
 
 	if (namedImport.getParent().getParent().getNamedImports().length === 1) {
@@ -150,7 +150,7 @@ export function transform_code(code, _is_ts, file_path) {
 function remove_throws(source) {
 	const logger = log_on_ts_modification(
 		source,
-		'Removed `throw` from redirect/error functions: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#redirect-and-error-are-no-longer-thrown-by-you'
+		'Removed `throw` from redirect/error functions: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#redirect-and-error-are-no-longer-thrown-by-you'
 	);
 
 	/** @param {string} id */
@@ -200,7 +200,7 @@ function add_cookie_note(file_path, source) {
 
 	const logger = log_on_ts_modification(
 		source,
-		'Search codebase for `@migration` and manually add the `path` option to `cookies.set/delete/serialize` calls: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#path-is-now-a-required-option-for-cookies'
+		'Search codebase for `@migration` and manually add the `path` option to `cookies.set/delete/serialize` calls: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#path-is-now-a-required-option-for-cookies'
 	);
 
 	const calls = [];
@@ -279,7 +279,7 @@ function replace_resolve_path(source) {
 
 	const logger = log_on_ts_modification(
 		source,
-		'Replaced `resolvePath` with `resolveRoute`: https://kit.svelte.dev/docs/migrating-to-sveltekit-2#resolvePath-has-been-removed'
+		'Replaced `resolvePath` with `resolveRoute`: https://svelte.dev/docs/kit/migrating-to-sveltekit-2#resolvePath-has-been-removed'
 	);
 
 	const name_node = named_import.getNameNode();