+++ sites/kit.svelte.dev/src/routes/docs/+layout.server.js
@@ -1,9 +1,0 @@
-import { get_docs_list } from '$lib/server/docs/index.js';
-
-export const prerender = true;
-
-export async function load() {
-	return {
-		sections: get_docs_list()
-	};
-}