+++ sites/kit.svelte.dev/src/routes/docs/[slug]/+page.server.js
@@ -1,7 +1,0 @@
-import { get_parsed_docs } from '$lib/server/docs/index.js';
-
-export const prerender = true;
-
-export async function load({ params }) {
-	return { page: await get_parsed_docs(params.slug) };
-}