+++ packages/enhanced-img/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/enhanced-img
 
+## 0.3.10
+### Patch Changes
+
+
+- docs: update URLs for new svelte.dev site ([#12857](https://github.com/sveltejs/kit/pull/12857))
+
+## 0.3.9
+### Patch Changes
+
+
+- chore: upgrade svelte-parse-markup ([#12793](https://github.com/sveltejs/kit/pull/12793))
+
 ## 0.3.8
 ### Patch Changes