+++ packages/package/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/package
 
+## 2.3.7
+### Patch Changes
+
+
+- docs: update URLs for new svelte.dev site ([#12857](https://github.com/sveltejs/kit/pull/12857))
+
+## 2.3.6
+### Patch Changes
+
+
+- docs: demonstrate sv migrate over prior commands ([#12840](https://github.com/sveltejs/kit/pull/12840))
+
 ## 2.3.5
 ### Patch Changes